requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# OpenAI + LangGraph
openai==1.54.0
langgraph==0.2.50
langchain-core==0.3.20

# Alpaca
alpaca-py==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Data / Technical Analysis
pandas==2.2.3
numpy==2.1.3
ta==0.11.0

# HTTP
httpx==0.27.2
aiohttp==3.10.10

# Timezone
pytz==2024.2

# Scheduling / async
apscheduler==3.10.4

# Polygon
polygon-api-client==1.14.2
